+++ noodles-vcf/CHANGELOG.md
@@ -4,8 +4,10 @@
 
 ### Changed
 
-  * vcf/io/writer/record/info/field/value/character: Remove colon (`:`) from
-    encode set.
+  * vcf/io/writer/record/info/field/value: Remove colon (`:`) from encode set
+    ([#339]).
+
+[#339]: https://github.com/zaeleus/noodles/issues/339
 
 ## 0.81.0 - 2025-08-25